update product and delete product

+++ Server/src/main/java/demo/com/example/testserver/product/repository/ProductRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductRepository extends JpaRepository<Product, Integer>, JpaSpecificationExecutor<Product> {
+public interface ProductRepository extends JpaRepository<Product, Long>, JpaSpecificationExecutor<Product> { // Changed Integer to Long
     // JpaSpecificationExecutor is added to support dynamic queries using Specifications
 
     // Add custom query methods if needed, for example:
